Moved stand-alone plotting function from plots.py to plot_functions.py Renamed plots.py to posterior.py.
